getAllTestForClient
