+++ src/main/java/kg/peaksoft/bilingualb6/api/TestController.java
@@ -3,11 +3,9 @@ package kg.peaksoft.bilingualb6.api;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kg.peaksoft.bilingualb6.dto.request.TestRequest;
-import kg.peaksoft.bilingualb6.dto.response.TestResponseForClient;
 import kg.peaksoft.bilingualb6.dto.response.SimpleResponse;
 import kg.peaksoft.bilingualb6.dto.response.TestInnerPageResponse;
 import kg.peaksoft.bilingualb6.dto.response.TestResponse;
-import kg.peaksoft.bilingualb6.dto.response.TestResponseGetTestByIdForClient;
 import kg.peaksoft.bilingualb6.service.TestService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -38,14 +36,6 @@ public class TestController {
     @PreAuthorize("hasAuthority('ADMIN')")
     public TestInnerPageResponse getById(@PathVariable Long id) {
         return testService.getTestById(id);
-    }
-
-    @Operation(summary = "Get test by id for client",
-            description = "This endpoint returns test by id and their questions and options for CLIENT")
-    @GetMapping("/client{id}")
-    @PreAuthorize("hasAuthority('CLIENT')")
-    public TestResponseGetTestByIdForClient getByIdForClient(@PathVariable Long id) {
-        return testService.getTestByIdForClient(id);
     }
 
     @Operation(summary = "Create test", description = "Create new test by ADMIN")
